// biontec/src/app/model/cliente-model.ts
export interface Cliente_T {

  id_cliente: number;
  nome_cliente: string;
  inscricaoest: string;
  pessoa: string;
  cpf: string;
  cnpj: string;
  cep: string;
  logradouro: string;
  numero: number;
  bairro: string;
  cidade: string;
  estado: string;
  telefone: string;
  celular: string;
  zap: string;
}